+++ ethers-solc/tests/mocked.rs
@@ -1,0 +1,1 @@
+//! mocked project tests
